+++ pkg/transport/listener.go
@@ -30,9 +30,7 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
-	"syscall"
 	"time"
-	"golang.org/x/sys/unix"
 
 	"go.etcd.io/etcd/pkg/v3/fileutil"
 	"go.etcd.io/etcd/pkg/v3/tlsutil"
@@ -66,8 +64,8 @@ func wrapTLS(scheme string, tlsinfo *TLSInfo, l net.Listener) (net.Listener, err
 	return newTLSListener(l, tlsinfo, checkSAN)
 }
 
-// NewListenerWithConfig
-func NewListenerWithConfig(addr, scheme string, tlsinfo *TLSInfo, sopts SocketOpts) (net.Listener, error) {
+// NewListenerWithSocketOpts
+func NewListenerWithSocketOpts(addr, scheme string, tlsinfo *TLSInfo, sopts SocketOpts) (net.Listener, error) {
 	if scheme == "unix" || scheme == "unixs" {
 		// unix sockets via unix://laddr
 		return NewUnixListener(addr)
@@ -89,23 +87,6 @@ func newListenConfig(addr, scheme string, sopts SocketOpts)  (net.ListenConfig, 
 		lc.Control = sopts.Control
 	}
 	return lc, nil
-}
-
-type SocketOpts []func(network, addr string, conn syscall.RawConn) error
-
-func (sopts SocketOpts) Control(network, addr string, conn syscall.RawConn) error {
-	for _, s := range sopts {
-		if err := s(network, addr, conn); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func SetReuseAddress(network, address string, conn syscall.RawConn) error {
-	return conn.Control(func(fd uintptr) {
-		syscall.SetsockoptInt(int(fd), syscall.SOL_SOCKET, unix.SO_REUSEPORT, 1)
-	})
 }
 
 type TLSInfo struct {
