+++ pkg/transport/timeout_listener.go
@@ -32,7 +32,7 @@ func NewTimeoutListener(addr string, scheme string, tlsinfo *TLSInfo, rdtimeoutd
 	return newTimeoutListener(ln, addr, scheme, rdtimeoutd, wtimeoutd, tlsinfo)
 }
 
-func NewTimeoutListerWithConfig(addr string, scheme string, tlsinfo *TLSInfo, rdtimeoutd, wtimeoutd time.Duration, sopts SocketOpts ) (net.Listener, error) {
+func NewTimeoutListerWithSocketOpts(addr string, scheme string, tlsinfo *TLSInfo, rdtimeoutd, wtimeoutd time.Duration, sopts SocketOpts ) (net.Listener, error) {
 	config, err := newListenConfig(addr, scheme, sopts)
 	if err != nil {
 		return nil, err
