+++ pkg/transport/transport.go
@@ -15,7 +15,6 @@
 package transport
 
 import (
-
 	"net"
 	"net/http"
 	"strings"
@@ -70,4 +69,3 @@ func (urt *unixTransport) RoundTrip(req *http.Request) (*http.Response, error) {
 	req.URL.Scheme = strings.Replace(req.URL.Scheme, "unix", "http", 1)
 	return urt.Transport.RoundTrip(req)
 }
-
