+++ pkg/transport/unix_listener.go
@@ -15,11 +15,8 @@
 package transport
 
 import (
-	"golang.org/x/sys/unix"
-	"k8s.io/klog"
 	"net"
 	"os"
-	"syscall"
 )
 
 type unixListener struct{ net.Listener }
