+++ server/embed/etcd.go
@@ -459,7 +459,7 @@ func configurePeerListeners(cfg *Config) (peers []*peerListener, err error) {
 			}
 		}
 		peers[i] = &peerListener{close: func(context.Context) error { return nil }}
-		peers[i].Listener, err = rafthttp.NewListenerWithSocketOptions(u, &cfg.PeerTLSInfo, getSocketOptions(cfg))
+		peers[i].Listener, err = rafthttp.NewListenerWithSocketOpts(u, &cfg.PeerTLSInfo, cfg.GetSocketOptions())
 		if err != nil {
 			return nil, err
 		}
@@ -524,14 +524,6 @@ func (e *Etcd) servePeers() (err error) {
 	return nil
 }
 
-func getSocketOptions(cfg *Config) transport.SocketOpts {
-	ctls := transport.SocketOpts{}
-	if cfg.ReuseAddress {
-		ctls = append(ctls, transport.SetReuseAddress)
-	}
-	return ctls
-}
-
 func configureClientListeners(cfg *Config) (sctxs map[string]*serveCtx, err error) {
 	if err = updateCipherSuites(&cfg.ClientTLSInfo, cfg.CipherSuites); err != nil {
 		return nil, err
@@ -574,7 +566,7 @@ func configureClientListeners(cfg *Config) (sctxs map[string]*serveCtx, err erro
 			continue
 		}
 
-		if sctx.l, err = transport.NewListenerWithConfig(addr, u.Scheme, &cfg.ClientTLSInfo, getSocketOptions(cfg)); err != nil {
+		if sctx.l, err = transport.NewListenerWithSocketOpts(addr, u.Scheme, &cfg.ClientTLSInfo, cfg.GetSocketOptions()); err != nil {
 			return nil, err
 		}
 		// net.Listener will rewrite ipv4 0.0.0.0 to ipv6 [::], breaking
